- refine code
